refactor(omo-senpi): split oversized adapter tests
